refactor: decompose relay.ts into focused modules
Splits the 1,183-line relay.ts god module into 8 focused modules:

- relay/types.ts          — shared RelaySocket type alias
- relay/thinking-tracker.ts — thinking-block duration tracking (~65 lines)
- relay/ack-tracker.ts    — per-socket cumulative ack tracking (~30 lines)
- relay/push-tracker.ts   — push notification logic (~110 lines)
- relay/event-pipeline.ts — chunked assembly + main event handler (~220 lines)
- relay/messaging.ts      — session_message / session_trigger / trigger_response (~270 lines)
- relay/child-lifecycle.ts — cleanup_child_session / delink_children / delink_own_parent (~300 lines)
- relay/session-lifecycle.ts — register / session_end / exec_result / disconnect (~115 lines)
- relay/index.ts          — namespace entry point, wires all handlers (~50 lines)

Zero behavior changes. All relay-cleanup tests pass (18/18).
Updates viewer.ts and namespaces/index.ts import paths.
